Converter corrected & .appveyor.yml added

+++ src/js/Converter.js
@@ -4,7 +4,7 @@ export class ArrayBufferConverter {
   }
 
   toString() {
-    const view = new Uint8Array(this.loadedBuffer);
+    const view = new Uint16Array(this.loadedBuffer);
     let stringArray = view.slice();
     stringArray = Array.from(stringArray);
     const string = stringArray.map((item) => { item = String.fromCharCode(item); return item; }).join('');
